+++ crates/cojson-core-wasm/src/lib.rs
@@ -141,31 +141,33 @@ impl SessionLog {
     }
 
     /// Add an existing private transaction to the staging area.
-    /// The transaction is NOT committed until validateSignature() succeeds.
+    /// The transaction is NOT committed until commitTransactions() succeeds.
+    /// Note: made_at uses f64 because JavaScript's number type is f64.
     #[wasm_bindgen(js_name = addExistingPrivateTransaction)]
     pub fn add_existing_private_transaction(
         &mut self,
         encrypted_changes: String,
         key_used: String,
-        made_at: u64,
+        made_at: f64,
         meta: Option<String>,
     ) -> Result<(), CojsonCoreWasmError> {
         self.internal
-            .add_existing_private_transaction(encrypted_changes, key_used, made_at, meta)
+            .add_existing_private_transaction(encrypted_changes, key_used, made_at as u64, meta)
             .map_err(CojsonCoreWasmError::CoJson)
     }
 
     /// Add an existing trusting transaction to the staging area.
-    /// The transaction is NOT committed until validateSignature() succeeds.
+    /// The transaction is NOT committed until commitTransactions() succeeds.
+    /// Note: made_at uses f64 because JavaScript's number type is f64.
     #[wasm_bindgen(js_name = addExistingTrustingTransaction)]
     pub fn add_existing_trusting_transaction(
         &mut self,
         changes: String,
-        made_at: u64,
+        made_at: f64,
         meta: Option<String>,
     ) -> Result<(), CojsonCoreWasmError> {
         self.internal
-            .add_existing_trusting_transaction(changes, made_at, meta)
+            .add_existing_trusting_transaction(changes, made_at as u64, meta)
             .map_err(CojsonCoreWasmError::CoJson)
     }
